Rolling.jl Documentation

using Plots
using Rolling

x = range(1, 10; length=100)
y = sin.(x) + 0.5 * rand(length(x))

plot(x, y; label="raw", title="Rolling means")

for window in [5, 10, 20]
    state = FixedWindowAssociativeOp{Float64}(+, window; emit_early=false)

    z = []
    for value in y
        new_value = update_state!(state, value)
        if !isnothing(new_value)
            push!(z, new_value / window)
        else
            push!(z, NaN)
        end
    end

    plot!(x, z; label="mean $window", lw=2)
end
current()

Rolling.FixedWindowAssociativeOp — Type

FixedWindowAssociativeOp{T}

State necessary for accumulation over a rolling window of fixed size.

Fields

window_state::WindowedAssociativeOpState{T}: The underlying general-window state.

remaining_window::Int: How much of the window remains to be filled. Initially this will be set to the window size, and will then reduce for every value added until it reaches zero.

Rolling.FixedWindowAssociativeOp — Method

FixedWindowAssociativeOp{T}

Construct a new empty instance of FixedWindowAssociativeOp.

Arguments

op::Function: Any binary, associative, function.

window::Integer: The fixed window size.

Rolling.WindowedAssociativeOpState — Type

WindowedAssociativeOpState{T}

State associated with a windowed aggregation of a binary associative operator, in a numerically accurate fashion.

Wherever summation is discussed, we can consider any alternative binary, associative, operator. For example: +, *, max, min, &&, union

NB. It is interesting to observe that commutativity is not required by this algorithm, which is one of the reasons that it enjoys stable numerical performance.

Conceptually the window is maintained in two buffers:

    [---- A ---)[----- B ------)
        <                      >    <-- current window finishes at the end of B, and
                                        starts somewhere in A.

A is stored as a sequence of cumulative sums, such that as the "<" advances we merely pick out the correct element:

    x_i,   x_i-1 + x_i,  x_i-2 + x_i-1 + x_i

B is stored as both:

The sequence of values seen: x_i+1, x_i+2, x_i+3, ...

The total of that sequence: x_i+1 + x_i+2 + x_i+3 + ...

When the "<" advances from A to B, we discard A, and the subset of B remaining after < becomes the new A. In becoming A, we transform its representation into that of the cumulative sums. We create a new, empty, B.

O(1) amortized runtime complexity, and O(L) space complexity, where L is the typical window length.

Fields

op::Function: Any binary, associative, function.

previous_cumsum::Array{T, 1}: Corresponds to array A above.

ri_previous_cumsum::Int: A reverse index into previous_cumsum, once it contains values. It should be subtracted from end in order to obtain the appropriate index.

values::Array{T, 1}: Corresponds to array B above.

sum::Union{Nothing, T}: The sum of the elements in values.

Rolling.WindowedAssociativeOpState — Method

WindowedAssociativeOpState{T}

Create a new, empty, instance of WindowedAssociativeOpState.

Arguments

op::Function: Any binary, associative, function.

Returns

WindowedAssociativeOpState{T}: An empty instance.

Rolling.update_state! — Method

update_state!(
    state::FixedWindowAssociativeOp{T},
    value
)::Union{T, Nothing} where T

Add the specified value to the state. Drop a value from the window iff the window is full.

Returns

::FixedWindowAssociativeOp{T}: The instance state that was passed in.

Rolling.update_state! — Method

update_state!(
    state::WindowedAssociativeOpState{T},
    value,
    num_dropped_from_window::Integer
)::T where T

Add the specified value to the state, drop some number of elements from the start of the window, and return the aggregated quantity.

Arguments

state::WindowedAssociativeOpState{T}: The state to update (will be mutated).

value: The value to add to the end of the window - must be convertible to a T.

num_dropped_from_window::Integer: The number of elements to remove from the front of the window.

Returns

::WindowedAssociativeOpState{T}: The instance state that was passed in.

Rolling.window_full — Method

window_full(state::FixedWindowAssociativeOp)::Bool

Returns:

Bool: true iff the given state has a full window.

Rolling.window_size — Method

function window_size(state::WindowedAssociativeOpState{T})::Int where T

Get the current size of the window in state.

Arguments:

state::WindowedAssociativeOpState{T}: The state to query.

Returns:

Int: The current size of the window.

Rolling.window_value — Method

window_value(state::WindowedAssociativeOpState{T})::T where T

Get the value currently represented by the state.

Arguments:

state::WindowedAssociativeOpState{T}: The state to query.

Returns:

T: The result of aggregating over the values in the window.