Rolling.jl Documentation

Rolling.FixedWindowAssociativeOp — Type

FixedWindowAssociativeOp{T}

State necessary for accumulation over a rolling window of fixed size.

Fields

window_state::WindowedAssociativeOpState{T}: The underlying general-window state.

remaining_window::Int: How much of the window remains to be filled. Initially this will be set to the window size, and will then reduce for every value added until it reaches zero.

Rolling.FixedWindowAssociativeOp — Method

FixedWindowAssociativeOp{T}

Construct a new empty instance of FixedWindowAssociativeOp.

Arguments

op::Function: Any binary, associative, function.

window::Integer: The fixed window size.

emit_early::Bool: Iff true, compute a value before the window is full.

Rolling.WindowedAssociativeOpState — Type

WindowedAssociativeOpState{T}

State associated with a windowed aggregation of a binary associative operator, in a numerically accurate fashion.

Wherever summation is discussed, we can consider any alternative binary, associative, operator. For example: +, *, max, min, &&, union

NB. It is interesting to observe that commutativity is not required by this algorithm, which is one of the reasons that it enjoys stable numerical performance.

Conceptually the window is maintained in two buffers:

    [---- A ---)[----- B ------)
        <                      >    <-- current window finishes at the end of B, and
                                        starts somewhere in A.

A is stored as a sequence of cumulative sums, such that as the "<" advances we merely pick out the correct element:

    x_i,   x_i-1 + x_i,  x_i-2 + x_i-1 + x_i

B is stored as both:

The sequence of values seen: x_i+1, x_i+2, x_i+3, ...

The total of that sequence: x_i+1 + x_i+2 + x_i+3 + ...

When the "<" advances from A to B, we discard A, and the subset of B remaining after < becomes the new A. In becoming A, we transform its representation into that of the cumulative sums. We create a new, empty, B.

O(1) amortized runtime complexity, and O(L) space complexity, where L is the typical window length.

Fields

value_count::Int: The size of the window on the last update.

op::Function: Any binary, associative, function.

previous_cumsum::Array{T, 1}: Corresponds to array A above.

ri_previous_cumsum::Int: A reverse index into previous_cumsum, once it contains values. It should be subtracted from end in order to obtain the appropriate index.

values::Array{T, 1}: Corresponds to array B above.

sum::Union{Nothing, T}: The sum of the elements in values.

Rolling.WindowedAssociativeOpState — Method

WindowedAssociativeOpState{T}

Create a new, empty, instance of WindowedAssociativeOpState.

Arguments

op::Function: Any binary, associative, function.

Returns

WindowedAssociativeOpState{T}: An empty instance.

Rolling.update_state! — Method

update_state!(
    state::FixedWindowAssociativeOp{T},
    value
)::Union{T, Nothing} where T